moved SimpleEventQueue
